Clean up font experiment
