+++ src/backends/index.ts
@@ -1,0 +1,2 @@
+export { default as CanvasBackend } from "./CanvasBackend"
+export { default as SvgBackend } from "./SvgBackend"
