+++ src/index.ts
@@ -1,27 +1,32 @@
+import { CanvasBackend, SvgBackend } from "./backends"
+import layoutRoot, { Component, Settings } from "./layout"
+import render from "./render"
 
-import treeToSVG from "./svg/tree-to-svg"
-import treeToCanvas from "./canvas"
-import layoutRoot, {Component, Settings} from './layout/'
-
-const {registerFont: registerCanvasFont} = require('canvas')
-export {registerCanvasFont}
-
-export {Component, Settings}
-
-const defaultSettings = {width: 500, height: 500, fontCache: {fonts: {}, fallbacks: {}}, basePath: "./", renderPath: './', assetMap: {}}
-
-export const renderToSVGString = (root: Component, settings: Settings = defaultSettings) => {
-    if (!settings.fontCache)  {
-        throw new Error('No font cache provided')
-    }
-    return treeToSVG(layoutRoot(root, settings), settings)
+const defaultSettings: Settings = {
+  width: 500,
+  height: 500,
+  fontCache: { fonts: {}, fallbacks: {} },
+  basePath: "./",
+  renderPath: "./",
+  assetMap: {}
 }
 
-export const renderToCanvas = async (dest: string, root: Component, settings: Settings = defaultSettings) => {
-    if (!settings.fontCache)  {
-        throw new Error('No font cache provided')
-    }
-    await treeToCanvas(dest, layoutRoot(root, settings), settings)
+export const renderToSvg = async (
+  root,
+  settings: Settings = defaultSettings
+) => {
+  const formattedRoot = layoutRoot(root, settings)
+  const backend = new SvgBackend(formattedRoot)
+  await render(backend, formattedRoot, settings)
+  return String(backend)
 }
 
-export {initFontCache, loadFont, addFontFallback} from './layout/font-utils'
+export const renderToCanvas = async (
+  ctx,
+  root,
+  settings: Settings = defaultSettings
+) => {
+  const formattedRoot = layoutRoot(root, settings)
+  const backend = new CanvasBackend(ctx)
+  await render(backend, formattedRoot, settings)
+}
