+++ src/layout/extract-text.ts
@@ -1,14 +1,11 @@
-export interface AttributedStyle { start: number, end: number, style: any }
+import { AttributedStyle, TextWithAttributedStyle } from "../text-layout"
 
-export interface TextWithAttributedStyle { text: string, attributedStyles: AttributedStyle[] }
-
-const flattenStyles = (styles): object => Array.isArray(styles)
-  ? Object.assign({}, ...styles)
-  : (styles || {})
+const flattenStyles = (styles): object =>
+  Array.isArray(styles) ? Object.assign({}, ...styles) : styles || {}
 
 const getStyles = component => flattenStyles(component.props.style)
 
-const mergeStyles = (a, b) => Object.keys(b).length > 0 ? { ...a, ...b } : a
+const mergeStyles = (a, b) => (Object.keys(b).length > 0 ? { ...a, ...b } : a)
 
 const defaultStyles = {
   color: "black",
@@ -17,7 +14,7 @@ const defaultStyles = {
   fontStyle: "normal",
   fontWeight: "normal",
   lineHeight: 18,
-  textAlign: "left",
+  textAlign: "left"
 }
 
 const appendStyleTo = (
@@ -25,9 +22,10 @@ const appendStyleTo = (
   text: string,
   style: object
 ) => {
-  const lastAttributedStyle = attributedStyles.length > 0
-    ? attributedStyles[attributedStyles.length - 1]
-    : null
+  const lastAttributedStyle =
+    attributedStyles.length > 0
+      ? attributedStyles[attributedStyles.length - 1]
+      : null
 
   if (lastAttributedStyle !== null && style === lastAttributedStyle) {
     lastAttributedStyle.end += text.length
@@ -43,7 +41,7 @@ export default (component): TextWithAttributedStyle => {
   const attributedStyles: AttributedStyle[] = []
 
   const iterate = (c, style = mergeStyles(defaultStyles, getStyles(c))) => {
-    (c.children || []).forEach(child => {
+    ;(c.children || []).forEach(child => {
       if (child == null) {
         /* Do nothing */
       } else if (typeof child !== "object") {
