+++ src/render/__tests__/__image_snapshots__/index-test-tsx-render-to-canvas-1-snap.png
