+++ src/render/rect.ts
